Added image preloading system
This implements a system for simplifying the preloading of images.

It does *not* attempt to address more complicated issues of preloading
images (i.e. would it be possible to dynamically load inline post
images) and it does not attempt to automatically find and load anything:
it only creates a framework for manually defining them. Which images are
to be loaded is left entirely up to the developer.